config: switch from toml to tomli
tomli library is up-to-date with TOML spec, whereas toml only supports
up to 0.5. This is an issue since I want to support heterogenous arrays
in the configuration, which the old toml doesn't support.

In addition, tomli is shipped with Python 3.11 as "tomllib", which means
newer Python versions don't even need the additional dependency.