violations: refactor violations to use rule ID enum
This refactoring gets rid of the weird violation factory-function method
of enumerating violation types, in favor of a new `Rule` enum that lists
all violation types. The checkers now report violations by instantiating
the `Violation` base class and providing an ID of the enum type to the
constructor.

This has lots of nice effects while still achieving the goal of the old
method, which was to have an actual symbol to reference IDs (that can be
checked by linters) as opposed to using strings. A key benefit is that
bugs like the previous commit are no longer possible, since we can
easily enumerate all violations. In addition, we don't need to import a
bunch of individual violation types, just the enum. This will also make
type-checking more useful once that's enabled.