parser: remove command parsing arg count assertion
The current implementation of utils.eval() can actually violate this
assertion in the case where you have multiple barewords after an eval
(e.g. `eval command arg1 arg`). In order to support functional checks of
these sorts of eval blocks, this commit removes this check to preserve
the behavior of utils.eval().

In a future commit, the utils.eval() implementation will be updated to
ensure that style checks are correct.

In other cases we still want this to be an invariant, but we can make
this easier to maintain by eventually moving command definitions to
static data, eliminating possible points of failure in the
implementation.