format: switch to hanging indents across the board
This commit is a big one, and may have a controversial impact on
tclfmt's style.

This commit removes the block indentation style that was previously used
in command substitutions and expressions. Some examples of the old
style:

```
set result [command \
               arg]
```

or

```
if { $foo &&
     $bar } {
    # ...
}
```

The new style enforces two universal properties:

1) All continuations are indented one level, regardless of the contents
   of the previous line.

That means the first example now looks like

```
set result [command \
    arg]
```

2) If braced expressions, parenthesized expressions, or function
   arguments span multiple lines, the enclosing braces/parens are set on
   different lines from the contents.

This makes the second example look like

```
if {
    $foo &&
    $bar
} {
    # ...
}
```

This helps make the code more readable when property 1 is being
enforced. Putting the closing bracket on its own line provides visual
separation between the end of a multi-line expression and body in
control flow commands, which is helpful since the expression is
indented to the same level as the body.

Putting the opening bracket on its own line lines up the start of all
lines in the expression.

This style is also consistent with the way lists and scripts work.

An additional benefit of this style change is that it simplifies the
tclfmt implementaiton. A very tricky bit has been facilitating both
hanging and block indents. The solution ended up being a bunch of hacks
that relied on hardcoding node types. Removing these hacks is nice, and
there's probably even more cleanup that can be done in follow-up
commits.

A final argument in favor of hanging indents is that it's a modern
style. Black, gofmt, and rustfmt all to indent in this style.