lexer: avoid slow reinitialization
This commit fixes a performance bug in the lexer. It turns out that each
call to `lex.lex()` performs a somewhat expensive reflection process to
generate a lexer table. This was being called every time the Lexer
object was initialized, which happens frequently, since the parser
instantiates a new lexer every time it performs a recursive parse of
command arguments.

In order to avoid this problem, this commit creates a singleton class
called `LexTable` that holds the lex token definitions and a single
initialized instance of a lexer. The Lexer constructor now just takes a
clone of this preinitialized lexer and resets its state.

This commit improves performance by ~17%, going from 13.3 to 11 seconds
to run tclint on the entire OpenROAD codebase.