Remove style checks from tclint!
This is a big change that reduces the size of the codebase quite a bit.
Since the project now has tclfmt, we can rely on the formatter as the
stylistic source of truth, rather than trying to check individual
violations using the linter. This will be much simpler going forward -
there were several lingering issues with the tclint style checks that
can now be closed.

User-facing things that are now gone:
- The style-specific rules. References will have to be removed from
  config files.
- No more --no-check-style flag, its behavior is now the default.
- No more style.allow-aligned-sets config setting, since tclfmt doesn't
  respect it.