main: rename tclint.py to main.py
This fixes an import issue when directly executing Python files in the
codebase that import the tclint package (e.g. python3
src/tclint/tclint.py).

When executing a file, the directory containing that file gets added to
the PYTHONPATH. That caused imports of tclint to resolve to the module
(tclint.py) rather than the package.

While executing these files directly isn't explicitly supported,
renaming this file has little cost (and main.py is arguably a better
name), and it makes the repo's behavior less suprising.