config: don't hardcode CWD as root for user-supplied config
This doesn't affect any existing code, since the tclint and tclfmt
entrypoints are the only `get_config` callers that supply a config
directly, and they supply the CWD as root anyways.

However, the LSP is going to be updated to to supply a direct path to
the config file, at which point this behavior will be required to avoid
a bug, since the LSP will supply the current workspace folder as the
root (not the CWD).