tests: check in OpenROAD command spec
A future commit is going to strip out the OpenROAD plugin support and
make its command spec format a first-class part of the codebase. This
gives us a test case for working on the command spec checking. We can
update it as the spec evolves via a migration script.

This spec was generated from OpenROAD v2.0-13652-g6fc686431.