cli: move main.py to cli/tclint.py
This makes the project organization more consistent. It also reflects
how the framing of the project is evolving to emphasize that it's a
collection of utilities, not just a linter.

This does reintroduce the import issue that was fixed by
6de1f809c6880909931a3b64a61fff7497befeee (but even worse, it breaks the
tclsp tests). However, I've discovered running with `python3 -I` solves
this problem.