commands: document some built-ins
I'm considering using the first non-blank docstring lines of functions
that define commands as help strings, in lieu of being able to generate
them (with a command spec).

I can finish this up once a relevant feature is immplemented.

Was also a good chance to add a link to the Tcl 8.6 docs. Google now
brings me to 9.0.