checks: recognize -nonewline as positional arg of puts
Currently, for this TCL code:
...
puts -nonewline "foo "
puts "bar"
...
we get:
...
unrecognized argument for puts: -nonewline [command-args]
...

As reported here ( https://github.com/nmoroze/tclint/issues/106 ), this is not
a TCL error, and tclint should not report this.

The "-nonewline" is listed in builtin.py as positional of puts.

The problem is that check_arg_spec decides that -nonewline is a switch rather
than a positional, based on how it looks (BareWord starting with "-").

Fix this in check_arg_spec by finding -nonewline in arg_spec["positionals"],
and deciding it's in fact a positional.