ci: add .pre-commit-config.yaml
We have a pre-commit script util/pre-commit calling various checkers as well
as pytest, but it's still up to the developer to call it.

Using the pre-commit framework [1], we can make sure that the checkers are
called every commit.

Add a file .pre-commit-config.yaml containing hooks for:
- black
- flake8
- mypy
- codespell

I could also have added pytest, but it seems excessive to do this for every
commit.

I've added pre-commit to the dev group of dependencies, which means the git
pre-commit checks will get activated for each developer that runs the setup
command.

It's possible to not do this, but then the changes to util/pre-commit will
have to be reverted.

Fixes issue:
- https://github.com/nmoroze/tclint/issues/112

[1] https://pre-commit.com/