ci: add .pre-commit-config.yaml
We have a pre-commit script util/pre-commit calling various checkers as well
as pytest, but it's still up to the developer to call it.

Using the pre-commit framework [1], we can make sure that the checkers are
called every commit.

Add a file .pre-commit-config.yaml containing hooks for:
- black
- flake8
- mypy
- codespell

I could also have added pytest, but it seems excessive to do this for every
commit.

To start using this setup, install the pre-commit hooks in the repository:
```
$ pre-commit install
pre-commit installed at .git/hooks/pre-commit
$
```
and the checkers will be called automatically when committing:
```
$ echo "# Necesary comment." >> README.md
$ git commit -a -m Comment
black...............................................(no files to check)Skipped
flake8..............................................(no files to check)Skipped
mypy................................................(no files to check)Skipped
codespell...............................................................Failed
- hook id: codespell
- exit code: 65

README.md:93: Necesary ==> Necessary

$
```

If the checkers fail, the commit is aborted.  A common situation is that a
developer wants to commit a WIP patch that fails one of the checks.  Use
`--no-verify` for that:
```
$ git commit -a -m Comment --no-verify
[pre-commit-config-yaml e640915] Comment
 1 file changed, 1 insertion(+)
$
```

More fine-grained control is available using the SKIP environment variable.

The setup can be simply removed using:
```
$ pre-commit uninstall
pre-commit uninstalled
$
```

Fixes issue:
- https://github.com/nmoroze/tclint/issues/112

[1] https://pre-commit.com/