checking: improve error message for args checking of if builtin
Currently error messages for all malformed ifs are identical:
```
$ echo "if" | tclint -
(stdin):1:1: error parsing command arguments, possibly malformed if command \
  [command-args]
$ echo "if 1" | tclint -
(stdin):1:1: error parsing command arguments, possibly malformed if command \
  [command-args]
```

Try to make the errors easier to understand by making them specific (dropped
repeating prefix and postfix for readability):
```
$ echo "if" | tclint -
Expected condition argument in 'if'
$ echo "if 1" | tclint -
Expected then or body argument in 'if'
$ echo "if 1 then" | tclint -
Expected body argument in 'if'
$ echo "if 1 {} else" | tclint -
Expected body argument to 'else' part of 'if'
$ echo "if 1 {} else {} {}" | tclint -
Argument after complete 'if': BracedWord()
$ echo "if 1 {} elseif" | tclint -
Expected condition argument in 'elseif' part of 'if'
$ echo "if 1 {} elseif 2" | tclint -
Expected then or body argument in 'elseif' part of 'if'
$ echo "if 1 {} elseif 2 then" | tclint -
Expected body argument in 'elseif' part of 'if'
```

For an elsif typo though we have:
```
$ echo "if 1 {} elsif 2 {}" | tclint -
Argument after complete 'if': BareWord('2')
```
because elsif is parsed as else body.  Unfortunately this error message is
still not easy to understand.