plugins: support tclint --commands dynamic-plugin.py
Tclint supports plugins, both static and dynamic ones.

A static plugin can be specified using a JSON file:
```
$ tclint --commands foo.json ...
```

A dynamic plugin can be specified using a plugin discovery mechanism called
entry points, which requires wrapping the plugin in a python package and
advertising the plugin in the package metadata.

Add a simpler way of specifying a dynamic plugin: specifying the python file
containing it:
```
$ tclint --commands foo.py ...
```

Fixes issue:
- https://github.com/nmoroze/tclint/issues/121