lsp: refactor to store all configs by root
We're moving towards resolving relative directories in the config at
load time. For the LSP, in most cases we want to resolve a config from a
workspace root. That means that if a global config applies, it must be
loaded separately for each workspace.

The other case is when we have a file outside of a workspace root. The
LSP's current behavior is to use the file's parent directory for
resolving exclude patterns. To preserve this sort of behavior, we now
load configs on the fly relative to the root of a non-workspace file if
one is discovered.

This might suffer a slight performance hit on the first call to
diagnostics/formatting for a new non-workspace file, and also may result
in high memory utilization if many non-workspace files are open.
However, I expect most LSP users are within a workspace, so hopefully
this is not an issue.