config: discover config by walking up from source
This commit is a major change to tclint's config file discovery
mechanism. tclint (and tclfmt) now discover the config for a given
source file by walking the filesystem up from the source until it finds
a config file. This means that configs can be discovered even if tclint
is run from a subdirectory of a repository. It also means a project can
support different configs depending on directory. The LSP has been
updated with similar behavior, but it stops searching at workspace root
boundaries.

This aligns with what many other tools do, e.g. Ruff[^1].

This addresses #30 and #120, and will facilitate simplifying the config
file spec and implementation by dropping the fileset feature.

[^1]: https://docs.astral.sh/ruff/configuration/#config-file-discovery